Refactor components into common location
