+++ ml-agents/mlagents/trainers/components/curiosity/model.py
