+++ ml-agents/mlagents/trainers/components/entropy/signal.py
@@ -1,4 +1,4 @@
-from mlagents.trainers.ppo.components import RewardSignal
+from mlagents.trainers.components import RewardSignal
 from mlagents.trainers.policy import Policy
