+++ ml-agents/mlagents/trainers/components/gail/model.py
