+++ ml-agents/mlagents/trainers/ppo/__init__.py
@@ -1,4 +1,3 @@
 from .models import *
 from .trainer import *
 from .policy import *
-from .pre_training import *
